outputs/wandb/run-20260413_220014-pqdih2q3/files/requirements.txt
annotated-types==0.7.0
sympy==1.14.0
depth_anything_v2==0.0.0
certifi==2025.8.3
numpy==1.26.4
six==1.17.0
triton==3.3.1
aiosignal==1.4.0
torchdyn==1.0.6
jupyter_core==5.8.1
nvidia-curand-cu12==10.3.7.77
nvidia-cusparse-cu12==12.5.4.2
platformdirs==4.3.8
widgetsnbextension==4.0.14
poethepoet==0.10.0
torchvision==0.22.1
pyparsing==3.2.3
parso==0.8.4
psutil==7.0.0
opencv-python==4.11.0.86
zipp==3.23.0
wcwidth==0.2.13
packaging==25.0
diffusers==0.11.1
torchcfm==1.0.7
propcache==0.3.2
tzdata==2025.2
joblib==1.5.1
python-dateutil==2.9.0.post0
setuptools==80.9.0
ipython==8.37.0
efficientnet_pytorch==0.7.1
executing==2.2.0
scipy==1.15.3
nvidia-cuda-runtime-cu12==12.6.77
regex==2025.7.34
einops==0.8.1
charset-normalizer==3.4.3
multidict==6.6.4
MarkupSafe==3.0.2
POT==0.9.5
pandas==2.3.1
pip==26.0.1
urllib3==2.5.0
mpmath==1.3.0
ptyprocess==0.7.0
nvidia-cudnn-cu12==9.5.1.17
typing_extensions==4.14.1
lmdb==1.7.3
torchcde==0.2.5
decorator==5.2.1
matplotlib-inline==0.1.7
idna==3.10
torch==2.7.1
pyzmq==27.0.1
protobuf==6.31.1
Pygments==2.19.2
typing-inspection==0.4.1
jedi==0.19.2
contourpy==1.3.2
nvidia-cuda-cupti-cu12==12.6.80
frozenlist==1.7.0
pexpect==4.9.0
matplotlib==3.10.5
pure_eval==0.2.3
nvidia-nccl-cu12==2.26.2
wheel==0.46.3
pytorch-lightning==2.5.2
asttokens==3.0.0
nvidia-nvtx-cu12==12.6.77
threadpoolctl==3.6.0
debugpy==1.8.16
wandb==0.21.1
GitPython==3.1.45
trampoline==0.1.2
lightning-utilities==0.15.2
filelock==3.18.0
nvidia-cublas-cu12==12.6.4.1
pytz==2025.2
cycler==0.12.1
nvidia-cuda-nvrtc-cu12==12.6.77
absl-py==2.3.1
nvidia-cusparselt-cu12==0.6.3
huggingface-hub==0.25.2
cv-bridge==1.13.0.post0
warmup_scheduler==0.3
click==8.2.1
stack-data==0.6.3
pastel==0.2.1
nvidia-cusolver-cu12==11.7.1.2
torchdiffeq==0.2.5
Jinja2==3.1.6
torchsde==0.2.6
jupyterlab_widgets==3.0.15
sentry-sdk==2.34.1
scikit-learn==1.7.1
fsspec==2025.7.0
yarl==1.20.1
attrs==25.3.0
smmap==5.0.2
PyYAML==6.0.2
fonttools==4.59.0
aiohappyeyeballs==2.6.1
torchmetrics==1.8.1
ipywidgets==8.1.7
nvidia-nvjitlink-cu12==12.6.85
gitdb==4.0.12
networkx==3.4.2
aiohttp==3.12.15
traitlets==5.14.3
jupyter_client==8.6.3
pydantic_core==2.33.2
tornado==6.5.2
nvidia-cufft-cu12==11.3.0.4
ipykernel==6.30.1
tqdm==4.67.1
pillow==11.3.0
kiwisolver==1.4.9
prompt_toolkit==3.0.51
nvidia-cufile-cu12==1.11.1.6
comm==0.2.3
requests==2.32.4
nest-asyncio==1.6.0
tomlkit==0.13.3
importlib_metadata==8.7.0
pydantic==2.11.7
diffusion_policy==0.0.0
wheel==0.45.1
jaraco.functools==4.0.1
more-itertools==10.3.0
autocommand==2.2.2
zipp==3.19.2
platformdirs==4.2.2
backports.tarfile==1.2.0
inflect==7.3.1
typing_extensions==4.12.2
jaraco.collections==5.1.0
jaraco.text==3.12.1
tomli==2.0.1
importlib_metadata==8.0.0
packaging==24.2
jaraco.context==5.3.0
typeguard==4.3.0
